[refactor] refactor application service
