+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/controller/ApplicationController.java
@@ -35,7 +35,12 @@ import org.apache.streampark.console.core.enums.PermissionType;
 import org.apache.streampark.console.core.service.AppBuildPipeService;
 import org.apache.streampark.console.core.service.ApplicationBackUpService;
 import org.apache.streampark.console.core.service.ApplicationLogService;
-import org.apache.streampark.console.core.service.ApplicationService;
+import org.apache.streampark.console.core.service.application.ApplicationService;
+import org.apache.streampark.console.core.service.application.OpApplicationInfoService;
+import org.apache.streampark.console.core.service.application.QueryApplicationInfoService;
+import org.apache.streampark.console.core.service.application.ValidateApplicationService;
+import org.apache.streampark.console.core.service.application.deploy.K8sApplicationService;
+import org.apache.streampark.console.core.service.application.deploy.YarnApplicationService;
 import org.apache.streampark.flink.packer.pipeline.PipelineStatus;
 
 import org.apache.shiro.authz.annotation.RequiresPermissions;
@@ -50,6 +55,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -72,7 +78,18 @@ import java.util.stream.Collectors;
 @RequestMapping("flink/app")
 public class ApplicationController {
 
-  @Autowired private ApplicationService applicationService;
+  @Autowired
+  @Qualifier("streamApplicationService")
+  private ApplicationService applicationService;
+
+  @Autowired private OpApplicationInfoService applicationInfoService;
+
+  @Autowired private QueryApplicationInfoService queryApplicationInfoService;
+
+  @Autowired private ValidateApplicationService validateApplicationService;
+
+  @Autowired private K8sApplicationService k8sApplicationService;
+  @Autowired private YarnApplicationService yarnApplicationService;
 
   @Autowired private ApplicationBackUpService backUpService;
 
@@ -85,7 +102,7 @@ public class ApplicationController {
   @PostMapping("get")
   @RequiresPermissions("app:detail")
   public RestResponse get(Application app) {
-    Application application = applicationService.getApp(app);
+    Application application = queryApplicationInfoService.getApp(app);
     return RestResponse.success(application);
   }
 
@@ -95,7 +112,7 @@ public class ApplicationController {
   @PostMapping("create")
   @RequiresPermissions("app:create")
   public RestResponse create(Application app) throws IOException {
-    boolean saved = applicationService.create(app);
+    boolean saved = applicationInfoService.create(app);
     return RestResponse.success(saved);
   }
 
@@ -121,7 +138,7 @@ public class ApplicationController {
   @PostMapping(value = "copy")
   @RequiresPermissions("app:copy")
   public RestResponse copy(@Parameter(hidden = true) Application app) throws IOException {
-    Long id = applicationService.copy(app);
+    Long id = applicationInfoService.copy(app);
     Map<String, String> data = new HashMap<>();
     data.put("id", Long.toString(id));
     return id.equals(0L)
@@ -135,14 +152,14 @@ public class ApplicationController {
   @PostMapping("update")
   @RequiresPermissions("app:update")
   public RestResponse update(Application app) {
-    applicationService.update(app);
+    applicationInfoService.update(app);
     return RestResponse.success(true);
   }
 
   @Operation(summary = "Get applications dashboard data")
   @PostMapping("dashboard")
   public RestResponse dashboard(Long teamId) {
-    Map<String, Serializable> map = applicationService.dashboard(teamId);
+    Map<String, Serializable> map = queryApplicationInfoService.dashboard(teamId);
     return RestResponse.success(map);
   }
 
@@ -151,7 +168,7 @@ public class ApplicationController {
   @PostMapping("list")
   @RequiresPermissions("app:view")
   public RestResponse list(Application app, RestRequest request) {
-    IPage<Application> applicationList = applicationService.page(app, request);
+    IPage<Application> applicationList = queryApplicationInfoService.page(app, request);
     List<Application> appRecords = applicationList.getRecords();
     List<Long> appIds = appRecords.stream().map(Application::getId).collect(Collectors.toList());
     Map<Long, PipelineStatus> pipeStates = appBuildPipeService.listPipelineStatus(appIds);
@@ -188,7 +205,7 @@ public class ApplicationController {
   @PostMapping("mapping")
   @RequiresPermissions("app:mapping")
   public RestResponse mapping(Application app) {
-    boolean flag = applicationService.mapping(app);
+    boolean flag = applicationInfoService.mapping(app);
     return RestResponse.success(flag);
   }
 
@@ -198,7 +215,7 @@ public class ApplicationController {
   @PostMapping("revoke")
   @RequiresPermissions("app:release")
   public RestResponse revoke(Application app) {
-    applicationService.revoke(app);
+    applicationInfoService.revoke(app);
     return RestResponse.success();
   }
 
@@ -238,7 +255,7 @@ public class ApplicationController {
   @RequiresPermissions("app:start")
   public RestResponse start(@Parameter(hidden = true) Application app) {
     try {
-      applicationService.checkEnv(app);
+      validateApplicationService.checkEnv(app);
       applicationService.start(app, false);
       return RestResponse.success(true);
     } catch (Exception e) {
@@ -293,7 +310,7 @@ public class ApplicationController {
   @PostMapping("clean")
   @RequiresPermissions("app:clean")
   public RestResponse clean(Application app) {
-    applicationService.clean(app);
+    applicationInfoService.clean(app);
     return RestResponse.success(true);
   }
 
@@ -303,7 +320,7 @@ public class ApplicationController {
   @PostMapping("forcedStop")
   @RequiresPermissions("app:cancel")
   public RestResponse forcedStop(Application app) {
-    applicationService.forcedStop(app);
+    applicationInfoService.forcedStop(app);
     return RestResponse.success();
   }
 
@@ -316,28 +333,28 @@ public class ApplicationController {
   @Operation(summary = "Get application on yarn name")
   @PostMapping("name")
   public RestResponse yarnName(Application app) {
-    String yarnName = applicationService.getYarnName(app);
+    String yarnName = yarnApplicationService.getYarnName(app);
     return RestResponse.success(yarnName);
   }
 
   @Operation(summary = "Check the application exist status")
   @PostMapping("checkName")
   public RestResponse checkName(Application app) {
-    AppExistsState exists = applicationService.checkExists(app);
+    AppExistsState exists = queryApplicationInfoService.checkExists(app);
     return RestResponse.success(exists.get());
   }
 
   @Operation(summary = "Get application conf")
   @PostMapping("readConf")
   public RestResponse readConf(Application app) throws IOException {
-    String config = applicationService.readConf(app);
+    String config = queryApplicationInfoService.readConf(app);
     return RestResponse.success(config);
   }
 
   @Operation(summary = "Get application main-class")
   @PostMapping("main")
   public RestResponse getMain(Application application) {
-    String mainClass = applicationService.getMain(application);
+    String mainClass = queryApplicationInfoService.getMain(application);
     return RestResponse.success(mainClass);
   }
 
@@ -369,7 +386,7 @@ public class ApplicationController {
   @PostMapping("delete")
   @RequiresPermissions("app:delete")
   public RestResponse delete(Application app) throws InternalException {
-    Boolean deleted = applicationService.delete(app);
+    Boolean deleted = applicationInfoService.delete(app);
     return RestResponse.success(deleted);
   }
 
@@ -397,7 +414,7 @@ public class ApplicationController {
   @PostMapping("upload")
   @RequiresPermissions("app:create")
   public RestResponse upload(MultipartFile file) throws Exception {
-    String uploadPath = applicationService.upload(file);
+    String uploadPath = applicationInfoService.upload(file);
     return RestResponse.success(uploadPath);
   }
 
@@ -427,7 +444,7 @@ public class ApplicationController {
   @Operation(summary = "Check the application savepoint path")
   @PostMapping("checkSavepointPath")
   public RestResponse checkSavepointPath(Application app) throws Exception {
-    String error = applicationService.checkSavepointPath(app);
+    String error = queryApplicationInfoService.checkSavepointPath(app);
     if (error == null) {
       return RestResponse.success(true);
     } else {
@@ -458,7 +475,7 @@ public class ApplicationController {
   })
   @PostMapping(value = "k8sStartLog")
   public RestResponse k8sStartLog(Long id, Integer offset, Integer limit) throws Exception {
-    String resp = applicationService.k8sStartLog(id, offset, limit);
+    String resp = k8sApplicationService.k8sStartLog(id, offset, limit);
     return RestResponse.success(resp);
   }
 }
