+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/controller/ApplicationHistoryController.java
@@ -19,7 +19,8 @@ package org.apache.streampark.console.core.controller;
 
 import org.apache.streampark.common.enums.ExecutionMode;
 import org.apache.streampark.console.base.domain.RestResponse;
-import org.apache.streampark.console.core.service.ApplicationService;
+import org.apache.streampark.console.core.service.application.OpApplicationInfoService;
+import org.apache.streampark.console.core.service.application.deploy.K8sApplicationService;
 
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 
@@ -42,13 +43,14 @@ import java.util.List;
 @RequestMapping("flink/history")
 public class ApplicationHistoryController {
 
-  @Autowired private ApplicationService applicationService;
+  @Autowired private OpApplicationInfoService applicationInfoService;
+  @Autowired private K8sApplicationService k8sApplicationService;
 
   @Operation(summary = "List the upload jar history records")
   @PostMapping("uploadJars")
   @RequiresPermissions("app:create")
   public RestResponse listUploadJars() {
-    List<String> jars = applicationService.historyUploadJars();
+    List<String> jars = applicationInfoService.historyUploadJars();
     return RestResponse.success(jars);
   }
 
@@ -56,7 +58,7 @@ public class ApplicationHistoryController {
   @PostMapping("k8sNamespaces")
   @RequiresPermissions("app:create")
   public RestResponse listK8sNamespace() {
-    List<String> namespaces = applicationService.getRecentK8sNamespace();
+    List<String> namespaces = k8sApplicationService.getRecentK8sNamespace();
     return RestResponse.success(namespaces);
   }
 
@@ -69,7 +71,7 @@ public class ApplicationHistoryController {
       case KUBERNETES_NATIVE_SESSION:
       case YARN_SESSION:
       case REMOTE:
-        clusterIds = applicationService.getRecentK8sClusterId(executionMode);
+        clusterIds = k8sApplicationService.getRecentK8sClusterId(executionMode);
         break;
       default:
         clusterIds = new ArrayList<>(0);
@@ -82,7 +84,7 @@ public class ApplicationHistoryController {
   @PostMapping("flinkBaseImages")
   @RequiresPermissions("app:create")
   public RestResponse listFlinkBaseImage() {
-    List<String> images = applicationService.getRecentFlinkBaseImage();
+    List<String> images = k8sApplicationService.getRecentFlinkBaseImage();
     return RestResponse.success(images);
   }
 
@@ -90,7 +92,7 @@ public class ApplicationHistoryController {
   @PostMapping("flinkPodTemplates")
   @RequiresPermissions("app:create")
   public RestResponse listPodTemplate() {
-    List<String> templates = applicationService.getRecentK8sPodTemplate();
+    List<String> templates = k8sApplicationService.getRecentK8sPodTemplate();
     return RestResponse.success(templates);
   }
 
@@ -98,7 +100,7 @@ public class ApplicationHistoryController {
   @PostMapping("flinkJmPodTemplates")
   @RequiresPermissions("app:create")
   public RestResponse listJmPodTemplate() {
-    List<String> templates = applicationService.getRecentK8sJmPodTemplate();
+    List<String> templates = k8sApplicationService.getRecentK8sJmPodTemplate();
     return RestResponse.success(templates);
   }
 
@@ -106,7 +108,7 @@ public class ApplicationHistoryController {
   @PostMapping("flinkTmPodTemplates")
   @RequiresPermissions("app:create")
   public RestResponse listTmPodTemplate() {
-    List<String> templates = applicationService.getRecentK8sTmPodTemplate();
+    List<String> templates = k8sApplicationService.getRecentK8sTmPodTemplate();
     return RestResponse.success(templates);
   }
 }
