+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/task/FlinkRESTAPIWatcher.java
@@ -32,10 +32,12 @@ import org.apache.streampark.console.core.metrics.flink.CheckPoints;
 import org.apache.streampark.console.core.metrics.flink.JobsOverview;
 import org.apache.streampark.console.core.metrics.flink.Overview;
 import org.apache.streampark.console.core.metrics.yarn.YarnAppInfo;
-import org.apache.streampark.console.core.service.ApplicationService;
 import org.apache.streampark.console.core.service.FlinkClusterService;
 import org.apache.streampark.console.core.service.SavePointService;
 import org.apache.streampark.console.core.service.alert.AlertService;
+import org.apache.streampark.console.core.service.application.ApplicationService;
+import org.apache.streampark.console.core.service.application.OpApplicationInfoService;
+import org.apache.streampark.console.core.service.application.ValidateApplicationService;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.hc.client5.http.config.RequestConfig;
@@ -46,6 +48,7 @@ import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -70,7 +73,13 @@ import java.util.stream.Collectors;
 @Component
 public class FlinkRESTAPIWatcher {
 
-  @Autowired private ApplicationService applicationService;
+  @Autowired private ValidateApplicationService validateApplicationService;
+
+  @Autowired
+  @Qualifier("streamApplicationService")
+  private ApplicationService applicationService;
+
+  @Autowired private OpApplicationInfoService applicationInfoService;
 
   @Autowired private AlertService alertService;
 
@@ -155,7 +164,7 @@ public class FlinkRESTAPIWatcher {
   public void init() {
     WATCHING_APPS.clear();
     List<Application> applications =
-        applicationService.list(
+        validateApplicationService.list(
             new LambdaQueryWrapper<Application>()
                 .eq(Application::getTracking, 1)
                 .notIn(Application::getExecutionMode, ExecutionMode.getKubernetesMode()));
@@ -166,7 +175,7 @@ public class FlinkRESTAPIWatcher {
   public void doStop() {
     log.info(
         "FlinkRESTAPIWatcher StreamPark Console will be shutdown,persistent application to database.");
-    WATCHING_APPS.forEach((k, v) -> applicationService.persistMetrics(v));
+    WATCHING_APPS.forEach((k, v) -> applicationInfoService.persistMetrics(v));
   }
 
   /**
@@ -339,8 +348,7 @@ public class FlinkRESTAPIWatcher {
 
     // get overview info at the first start time
     if (STARTING_CACHE.getIfPresent(application.getId()) != null) {
-      FlinkCluster flinkCluster = getFlinkCluster(application);
-      Overview override = httpOverview(application, flinkCluster);
+      Overview override = httpOverview(application);
       if (override != null && override.getSlotsTotal() > 0) {
         application.setTotalTM(override.getTaskmanagers());
         application.setTotalSlot(override.getSlotsTotal());
@@ -387,7 +395,7 @@ public class FlinkRESTAPIWatcher {
               new LambdaUpdateWrapper<Application>()
                   .eq(Application::getId, application.getId())
                   .set(Application::getRelease, ReleaseState.DONE.get());
-          applicationService.update(updateWrapper);
+          validateApplicationService.update(updateWrapper);
           break;
         default:
           break;
@@ -421,7 +429,7 @@ public class FlinkRESTAPIWatcher {
     } else {
       WATCHING_APPS.put(application.getId(), application);
     }
-    applicationService.persistMetrics(application);
+    applicationInfoService.persistMetrics(application);
   }
 
   /**
@@ -451,7 +459,7 @@ public class FlinkRESTAPIWatcher {
             currentState.name());
         cleanSavepoint(application);
         application.setState(currentState.getValue());
-        if (StopFrom.NONE.equals(stopFrom) || applicationService.checkAlter(application)) {
+        if (StopFrom.NONE.equals(stopFrom) || validateApplicationService.checkAlter(application)) {
           if (StopFrom.NONE.equals(stopFrom)) {
             log.info(
                 "FlinkRESTAPIWatcher getFromFlinkRestApi, job cancel is not form StreamPark,savePoint expired!");
@@ -545,7 +553,7 @@ public class FlinkRESTAPIWatcher {
           if (flinkAppState.equals(FlinkAppState.FAILED)
               || flinkAppState.equals(FlinkAppState.LOST)
               || (flinkAppState.equals(FlinkAppState.CANCELED) && StopFrom.NONE.equals(stopFrom))
-              || applicationService.checkAlter(application)) {
+              || validateApplicationService.checkAlter(application)) {
             alertService.alert(application, flinkAppState);
             stopCanceledJob(application.getId());
             if (flinkAppState.equals(FlinkAppState.FAILED)) {
@@ -665,8 +673,7 @@ public class FlinkRESTAPIWatcher {
     return yarnRestRequest(reqURL, YarnAppInfo.class);
   }
 
-  private Overview httpOverview(Application application, FlinkCluster flinkCluster)
-      throws IOException {
+  private Overview httpOverview(Application application) throws IOException {
     String appId = application.getAppId();
     if (appId != null) {
       if (application.getExecutionModeEnum().equals(ExecutionMode.YARN_APPLICATION)
